+++ packages/indices/src/index.ts
@@ -1,0 +1,10 @@
+export {
+	getIndexAbove,
+	getIndexBelow,
+	getIndexBetween,
+	getIndices,
+	getIndicesAbove,
+	getIndicesBelow,
+	getIndicesBetween,
+	sortByIndex,
+} from './lib/reordering'
