+++ packages/utils/src/index.ts
@@ -24,4 +24,5 @@ export {
 	objectMapValues,
 } from './lib/object'
 export { rafThrottle, throttledRaf } from './lib/raf'
+export { sortById } from './lib/sort'
 export { isDefined, isNonNull, isNonNullish, structuredClone } from './lib/value'
